Add Exceptions/all test cases

+++ accountservice/src/main/java/com/demo/accountservice/repo/AccountRepo.java
@@ -3,12 +3,12 @@ package com.demo.accountservice.repo;
 import com.demo.accountservice.model.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.math.BigInteger;
 import java.util.List;
-import java.util.Optional;
 
 
 public interface AccountRepo extends JpaRepository<Account,Integer>{
-    public List<Account> findAccountByCustomerId(Integer id);
+    public List<Account> findAccountByCustomerId(BigInteger id);
 
 
 }
